+++ resources/views/fields/relationships/belongs-to.blade.php
@@ -1,3 +1,14 @@
+@if($element->isCreatable())
+{!! $element->createButton() !!}
+
+<x-moonshine::divider />
+
+@fragment($element->getRelationName())
+<div x-data="fragment('{{ $element->fragmentUrl() }}')"
+@defineEvent('fragment-updated', $element->getRelationName(), 'fragmentUpdate')
+>
+@endif
+
 <x-moonshine::form.select
     :attributes="$element->attributes()->merge([
         'id' => $element->id(),
@@ -12,3 +23,8 @@
     :asyncRoute="$element->isAsyncSearch() ? $element->asyncSearchUrl() : null"
 >
 </x-moonshine::form.select>
+
+@if($element->isCreatable())
+</div>
+@endfragment
+@endif
